Studio - Move hooks into src/apps/studio
Relocates the six remaining Studio hooks into src/apps/studio/hooks. Import paths only; no behaviour change.
