+++ src/apps/studio/hooks/useCrossModelConnectField.ts
@@ -3,9 +3,9 @@ import { useSelector } from "react-redux";
 import {
   useGetContentModelFieldsQuery,
   useGetContentModelsQuery,
-} from "../../../../../../../shell/services/instance";
-import { AppState } from "../../../../../../../shell/store/types";
-import { ConnectField } from "../../../../../../studio/hooks/studioTypes";
+} from "shell/services/instance";
+import { AppState } from "shell/store/types";
+import { ConnectField } from "./studioTypes";
 
 // Read-back for a cross-item Parsley reference.
 //
