+++ src/apps/studio/hooks/useLayoutReorderState.ts
@@ -1,6 +1,6 @@
 import { useCallback, useMemo, useRef, useState } from "react";
-import { notify } from "../../../../../../../shell/store/notifications";
-import { LayoutBreadcrumbItem } from "../../../../../../studio/hooks/studioTypes";
+import { notify } from "shell/store/notifications";
+import { LayoutBreadcrumbItem } from "./studioTypes";
 
 type LayoutStructureItem = {
   layoutId: string;
